Re-chunk badly chunked speaker lines
closes #26, fixing our incorrect assumption that *all* new speeches are
separated in the Word document by a line of whitespace.

Though most speeches are indeed separated by whitespace, counterexamples are
fairly common.  So you might see:

    Mr BOB SMITH: Blah blah

    Mrs JANE DOE: Blah blah (OK)
    Mr BOB SMITH: Blah blah (Not OK)

In this example, the 3rd para (Mr BOB SMITH's reply) will be grouped as part of
Mrs JANE DOE's speech.  This seems to be because when a speech starts on a new
page of the Word document, the scribes deliberately don't add a newline.

To disambiguate, we can simply check if the speech matches the name_regexp.  So
we let .break_paras() return not just True and False, but 0 (empty line) or an
incrementing counter.

(I would have used a closure for the counter, but Python's closures are
read-only, *sob*.  So I'm using external state and a method to update it, e.g.
.chunking_counter and .increment_chunking())