Add model to represent batches of questions.
Questions are supplied in papers which contain multiple questions.
See http://www.parliament.gov.za/live/content.php?Category_ID=236
for a list of such papers.

This model is to allow us to store both the external (from the table)
and internal (from headers and titles) metadata of those files. The
hope is that this will help us in assigning answers to questions.

There should be only one questionpaper per (year/issue) combination
so we may as well constrain this in the database.