Get spliting regex finding same number of questions as before.
This is done by starting to use a </b> rather than a colon as
our indication of the end of the intro.

We're now finding 104 questions, which is the same as in the
test output file.

However, there should be one more, question 720, which is getting
missed because of a unicode character in a name.

Strip out most <b> and <i> tags.