Add a new scraper for committee appearances that uses the new PMG API
The old scraper code was confusing and difficult to maintain, so it's
no bad thing to start anew since we're switching anyway from getting
committee appearances from scraping the PMG website to using the new
API which Code4SA have provided.

There's still some parsing to be done, since the body of each
committee proceedings event is still just HTML authored by PMG; this
structure also seems to have changed from when we were scraping the
site.

The code supporting the --save-json and --import-to-sayit is slightly
simplified, but otherwise similar to the old scraper; it's the
--scrape option that's been entirely rewritten.