As a stop-gap measure, allow blacklist of PopIt person IDs
There are known problems with speeches being misattributed to a
particular person (Kai-Lie Chen with PopIt ID core_person:5421).
We should get to the bottom of this failure of name resolution,
but for the moment we just want the committee scraper to run
without recreating these misattributions.

If we can't fix the name resolution issue, it would be better to add an
interface in the Django admin to let PMG blacklist people or appearances
as they spot them, so this is really an interim solution.