ZA Only associate committee appearances with current MPs.
Committee appearances will now only be associated with people
who were members of the national assembly on the date of the
committee meeting is question.

For committee meetings this should mean that we no longer see
1) modern members being associated with speeches before their time
(https://github.com/mysociety/pombola/issues/1859)
2) MPLs associated with National Assembly committee meetings
(https://github.com/mysociety/pombola/issues/1858)

This change also means that if a popit person is not found for
a speech that it will always have speaker net to NULL. Previously
the speaker would have been set to NULL if a popit person was
blacklisted. A new speaker would have been created with no popit
person if the popit person wasn't found.

This means that speeches which would once have had a Speaker
associated with no popit person will now just have no speaker
and will presumably look like narrative elements.

Previously the bits with no speaker at all were being given
a speaker with name '(narrative)' which doesn't seem right in
SayIt terms - I'd expect no speaker and a speech type of
'narrative'.

For committee appearances this is all entirely sensible. We display
nothing on the site, just link out to PMG, and there is no way to
actually view a speaker who isn't a popit person.

For plenary appearances we should always have an appropriate
popit_person too. If such a person isn't found, speeches which would
previously have appeared with a speaker with no popit person
attached will now appear as narrative.

In order to make this work I've added an optional acceptance check
for people in import_base's get_person method. This can be set as
a method on the importer, and defaults to returning True if not set.
Currently this is only used by the committee appearances code, but
it could be expanded to do things like run the blacklist (though I
hope the blacklist code can largely be removed as a result of the
main change here to only associate speeches with current members).