Ensure existing answers w/o questions get new questions
It's possible that we have an answer already in the database,
but it's not associated with any question.  If we get a question and
answer pair from the API, where the answer already exists in the
database but not the question, the question should be created and
the existing answer linked with it.  This commit adds a test to
check that.