Fix decoding errors: self.stdout.write needs a unicode or 7-bit ASCII
self.stdout.write tests to see if the message that's being written ends
with u'\n' (i.e. a unicode). It does that with .endswith,
which fails with a UnicodeDecodeError if the message is a str that
contains top bit set octets. It calls force_str on the message before
actually writing so it works fine when passed a unicode.